[CACE]: fix cscope delete and reset conflict

+++ plugin/cace.vim
@@ -132,6 +132,8 @@ let s:caceFinding = 0
 
 let s:caceRetryCnt = 3
 
+let s:caceSkipCscopeReset = 0
+
 let s:caceMsgSymbolDict = {
 			\ 'start' : 'CACE-S',
 			\ 'end' : 'CACE-E'
@@ -394,6 +396,7 @@ function! <SID>CACELstJobExitCb(job, status)
 			return
 		endif
 		call <SID>CACEProgressTrace(1)
+		let s:caceSkipCscopeReset = 1
 		call <SID>CACECleanDB('cscope')
 		call <SID>CACEStartJob('cscope')
 	else
@@ -603,16 +606,20 @@ function! <SID>CACELoadHLEDB()
 endfunction
 
 function! <SID>CACELoadCscopeDB()
-	" kill will disconnect db. but will leads find error. Use reset instead.
-	silent exe "cs reset"
+	if !s:caceSkipCscopeReset
+		" kill will disconnect db. but will leads find error. Use reset instead.
+		silent exe "cs reset"
+	endif
 	let db = findfile("cscope.out", ".;")
 	if (!empty(db))
 		let path = strpart(db, 0, match(db, "/cscope.out$"))
 		setlocal nocscopeverbose
 		exe "cs add " . db . " " . path
+		let s:caceSkipCscopeReset = 0
 		setlocal cscopeverbose
 	elseif $CSCOPE_DB != ""
 		exe "cs add " . $CSCOPE_DB
+		let s:caceSkipCscopeReset = 0
 	endif
 endfunction
